Unpack Flyway in a separate build step
Adding the .tar.gz file in one layer and unpacking it in a second layer adds a lot of bloat to the image size.

Moving the unpacking and permissions to a separate build step allows the final Flyway image to be much smaller (from 627MB to 447Mb on my machine).